% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{clean_data}
\alias{clean_data}
\title{Function to extracts erotic trials}
\usage{
clean_data(raw_data)
}
\arguments{
\item{raw_data}{dataframe, df containing all trials}
}
\value{
The output is the processed dataframe, that can be analyzed with
the analysis functions \code{\link{analysis_confirmatory}}, \code{\link{analysis_robustness}},
\code{\link{analysis_exploratory}}.
}
\description{
This function takes the raw data which contains all the trials
(see \code{\link{read_data}}), and excludes non-erotic trials, and
trials with missing trial numbers.
}
\examples{
\donttest{
tpp_processed_data <- clean_data(raw_data = example_m0)
}
}
