% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_confirmatory.R
\name{confirmatory_bayes_factor}
\alias{confirmatory_bayes_factor}
\title{Primary analysis: Bayes factors analysis}
\usage{
confirmatory_bayes_factor(success, total_n)
}
\arguments{
\item{success}{numeric, total number of successful guesses on erotic trials}

\item{total_n}{numeric, total number of all erotic trials}
}
\value{
The function returns a list of three numeric values with the rounded
Bayes factor calculated with each M1 beta prior.
}
\description{
The function computes the BF01 based on the successful guesses and the
total number of erotic trials. M0 is a point-value model denoting
that the chance of each successful guess is 50%. For M1 the function uses
three different beta priors on the probability of correct guesses:
knowledge-base prior, uniform prior, replication prior. In all cases we use
a one-tailed hypothesis testing (p > 0.5 and p < 1).
}
\examples{
\donttest{
# Including only erotic trials
tpp_processed_data <- clean_data(raw_data = example_m0)
# Running the confirmatory analysis
bayes_factor_result <- confirmatory_bayes_factor(success = sum(tpp_processed_data$sides_match, na.rm = T),
                                                 total_n = nrow(tpp_processed_data$sides_match))
# Checking the results
bayes_factor_results
}
}
\seealso{
Other confirmatory functions: 
\code{\link{confirmatory_mixed_effect}()}
}
\concept{confirmatory functions}
