% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_confirmatory.R
\name{confirmatory_mixed_effect}
\alias{confirmatory_mixed_effect}
\title{Primary analysis: Mixed-effects logistic regression}
\usage{
confirmatory_mixed_effect(processed_data, n_iteration = 1)
}
\arguments{
\item{processed_data}{a dataframe containing only erotic and non empty trials}

\item{n_iteration}{numeric, the nth time the analysis is conducted}
}
\value{
A list of three containing the width, the lower and the upper border
of the confidence interval of 95%.
}
\description{
This function calculates the mixed-effects logistic regression for
the primary analysis. A random intercept is placed on the participants
to predict the outcome of the guesses in erotic trials. Then the function
calculates the Wald confidence interval around the estimate of the odds
ratio of success and transforms it to logit probability.
}
\section{Note}{
 If the mixed-effects logistic regression is run by multiple
  times during the primary confirmatory analysis the alpha level is corrected
  by for each iteration.
}

\examples{
 
\donttest{
# Including only erotic trials
tpp_processed_data <- clean_data(raw_data = example_m0)
# Running the confirmatory analysis
mixed_effect_result <- confirmatory_mixed_effect(processed_data, n_iteration = 1)
# Checking the results
mixed_effect_result
}
}
\seealso{
Other confirmatory functions: 
\code{\link{confirmatory_bayes_factor}()}
}
\concept{confirmatory functions}
