% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_confirmatory.R
\name{cumulative_success}
\alias{cumulative_success}
\title{Calculating cumulative success}
\usage{
cumulative_success(df)
}
\arguments{
\item{df}{dataframe, the input dataframe}
}
\value{
The function returns a dataframe with the number
of trials in one column and the number of successful trials
in the other column.
}
\description{
This function calculates the number of trials and the
number of successful trials cumulatively. If raw data
are provided the function drops all the non-erotic
trials (see \code{\link{clean_data}}).
}
\examples{
\donttest{
# Get total number of trials and success cumulatively
cumulative_success_res <- cumulative_success(df = example_m0)
}
}
