% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_example_data.R
\name{generate_example_data}
\alias{generate_example_data}
\title{Generate example data}
\usage{
generate_example_data()
}
\description{
This code generates realistic example data for
the Transparent Psi Project.
}
