% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_example_data.R
\name{generate_session}
\alias{generate_session}
\title{Generate example session}
\usage{
generate_session(
  n_iteration,
  erotic_trial_size_per_participant,
  m0_prob,
  m1_prob,
  esp_user_percentage,
  chance_for_refuse_consent,
  chance_for_stopping_session
)
}
\arguments{
\item{erotic_trial_size_per_participant}{number of erotic trials performed per participant}

\item{m0_prob}{probability of successful guess if M0 is true}

\item{m1_prob}{probability of successful guess among ESP-users in the simulated sample}

\item{esp_user_percentage}{the percentage of ESP-users, or ESP-capable individuals in the population}

\item{chance_for_refuse_consent}{chance of refusing consent or being excluded due to one of the eligibility criteria (e.g. underage)}

\item{chance_for_stopping_session}{chance of stopping prematurely in each trial, to simulate some missing data due to unexpected events}
}
\value{
The function returns a dataframe containing the realistic
example data of one participant. The returned dataframe has 22 variables
and 1 to 41 rows depending on the results of the random probability
functions.
}
\description{
This function generates the example data of one participant.
}
