% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{get_links}
\alias{get_links}
\title{Get datafile download links from a Github folder}
\usage{
get_links(folder_path)
}
\arguments{
\item{folder_path}{character, url to the Github folder that contains the datafiles}
}
\value{
The function returns the download links as a character vector.
}
\description{
This function gets the download urls of every file in a Github
folder.
}
