% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference_confirmatory_bf}
\alias{inference_confirmatory_bf}
\title{Inference for Bayesian proportion tests}
\usage{
inference_confirmatory_bf(bf)
}
\arguments{
\item{bf}{numeric, the resulting Bayes factor}
}
\value{
The function returns a character value that is
either M1, M0 or Inconclusive.
}
\description{
This function make the inference for a Bayesian
proportion test based on the result of the test. If
the Bayes factor is lower than 1/25 the test supports
M1, but if the Bayes factor is higher than 25 the test
supports M0, otherwise inconclusive.
}
\seealso{
Other inference functions, confirmatory functions: 
\code{\link{inference_confirmatory_combined}()},
\code{\link{inference_confirmatory_mixed_effect}()}
}
\concept{inference functions, confirmatory functions}
