% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference_confirmatory_combined}
\alias{inference_confirmatory_combined}
\title{Inference for the confirmatory analysis}
\usage{
inference_confirmatory_combined(
  n_iteration,
  confirmatory_nhst_inference,
  confirmatory_bf_inference
)
}
\arguments{
\item{n_iteration}{numeric, the number of iterations for the mixed effect analysis}

\item{confirmatory_nhst_inference}{character, the inference based on the confirmatory NHST mixed analysis}

\item{confirmatory_bf_inference}{character, the inference based on the Bayes factor analysis with 3 priors}
}
\value{
The function returns a character value that is
either M1, M0, Inconclusive or Ongoing.
}
\description{
This function calculates the final inference for the
confirmatory analysis (see \code{\link{analysis_confirmatory}}).
If all the four primary tests support M0 or M1 the final inference
is conclusive. If after the last checkpoint not all the four
tests support M0 or M1 the inference is inconclusive, otherwise
the study is ongoing.
}
\seealso{
Other inference functions, confirmatory functions: 
\code{\link{inference_confirmatory_bf}()},
\code{\link{inference_confirmatory_mixed_effect}()}
}
\concept{inference functions, confirmatory functions}
