% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference_confirmatory_mixed_effect}
\alias{inference_confirmatory_mixed_effect}
\title{Inference for mixed effect test}
\usage{
inference_confirmatory_mixed_effect(mixed_ci_u, mixed_ci_l)
}
\arguments{
\item{mixed_ci_u}{numeric, the upper bound of the confidence interval}

\item{mixed_ci_l}{numeric, the lower bound of the confidence interval}
}
\value{
The function returns a character value that is
either M1, M0 or Inconclusive.
}
\description{
This function make the inference for the mixed-effects
logistic regression test. If the upper bound of the resulting
confidence interval is lower than 0.51, the test supports M0,
if the lower bound is higher than 0.5, the test supports M1.
Otherwise inconclusive.
}
\seealso{
Other inference functions, confirmatory functions: 
\code{\link{inference_confirmatory_bf}()},
\code{\link{inference_confirmatory_combined}()}
}
\concept{inference functions, confirmatory functions}
