% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference_robustness_bf}
\alias{inference_robustness_bf}
\title{Inference for robustness test using Bayesian parameter estimation}
\usage{
inference_robustness_bf(hdi_l, hdi_u)
}
\arguments{
\item{hdi_l}{numeric, the lower bound of the HDI}

\item{hdi_u}{numeric, the upper bound of the HDI}
}
\value{
The function returns a character that is either M0, M1 or Inconclusive.
}
\description{
This function makes the inference for the robustness
analysis test that uses Bayesian parameter estimation  of the
main parameter value (see \code{\link{robustness_bf_analysis}}).
}
\seealso{
Other inference functions, robustness functions: 
\code{\link{inference_robustness_combined}()},
\code{\link{inference_robustness_nhst}()}
}
\concept{inference functions, robustness functions}
