% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference_robustness_combined}
\alias{inference_robustness_combined}
\title{Inference for robustness analysis}
\usage{
inference_robustness_combined(
  confirmatory_analysis_inference,
  inference_robustness_nhst,
  inference_robustness_bayes_par_est
)
}
\arguments{
\item{confirmatory_analysis_inference}{character}

\item{inference_robustness_nhst}{character}

\item{inference_robustness_bayes_par_est}{character}
}
\value{
The function returns a character value.
}
\description{
The main analysis inference is only robust if all robustness
tests came to the same inference as the final inference
of the primary analysis (see \code{\link{inference_confirmatory_combined}}).
}
\seealso{
Other inference functions, robustness functions: 
\code{\link{inference_robustness_bf}()},
\code{\link{inference_robustness_nhst}()}
}
\concept{inference functions, robustness functions}
