% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference_robustness_nhst}
\alias{inference_robustness_nhst}
\title{Inference for robustness test using NHST}
\usage{
inference_robustness_nhst(equivalence_test_p, equality_test_p)
}
\arguments{
\item{equivalence_test_p}{numeric, the p-value of the equivalence test}

\item{equality_test_p}{numeric, the p-value of the equality test}
}
\value{
The function returns a character that is either M0, M1 or Inconclusive.
}
\description{
This function makes the inference for the robustness
analysis test that uses frequentist proportion test
(see \code{\link{robustness_nhst_analysis}}).
}
\seealso{
Other inference functions, robustness functions: 
\code{\link{inference_robustness_bf}()},
\code{\link{inference_robustness_combined}()}
}
\concept{inference functions, robustness functions}
