% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculation_helpers.R
\name{logit2prob}
\alias{logit2prob}
\title{Convert logit to probability}
\usage{
logit2prob(logit)
}
\arguments{
\item{logit}{numeric, the result of s logistic regression}
}
\description{
This is used for conversion of the results of the
logistic regression to the probability scale.
}
