% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculation_helpers.R
\name{mode_HDI}
\alias{mode_HDI}
\title{Function calculating the highest density interval using sampling}
\usage{
mode_HDI(scale, density, crit_width = 0.95, n_samples = 1e+05)
}
\description{
We use \code{\link[HDInterval]{hdi}} function in our
analysis. The present helper function is needed for
the Bayesian parameter estimation robustness test.
}
