% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_confirmatory.R
\name{plot_confirmatory}
\alias{plot_confirmatory}
\title{Plotting the results of the primary confirmatory analysis}
\usage{
plot_confirmatory(confirmatory_results)
}
\arguments{
\item{confirmatory_result}{tibble, output of the \code{\link{analyis_confirmatory}} function}
}
\value{
The function returns a ggplot object.
}
\description{
The function visualize the results of the confirmatory
analysis \code{\link{analyis_confirmatory}}. The Bayes factors
calculated at each new experimental trial are displayed on this figure.
}
\seealso{
Other plotting functions: 
\code{\link{plot_exploratory}()},
\code{\link{plot_robustness}()}
}
\concept{plotting functions}
