% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exploratory.R
\name{plot_exploratory}
\alias{plot_exploratory}
\title{Plotting the results of the exploratory analysis}
\usage{
plot_exploratory(
  success_rates_theoretical_prop,
  success_rates_empirical_prop,
  possible_success_rates
)
}
\arguments{
\item{success_rates_theoretical_prop}{numeric vector, output of the \code{\link{analysis_exploratory}} function}

\item{success_rates_empirical_prop}{numeric vector, output of the \code{\link{analysis_exploratory}} function}

\item{possible_success_rates}{numeric vector, output of the \code{\link{analysis_exploratory}} function}
}
\value{
The function returns a ggplot object.
}
\description{
The function visualize the results of the exploratory
analysis \code{\link{analysis_exploratory}}.
}
\seealso{
Other plotting functions: 
\code{\link{plot_confirmatory}()},
\code{\link{plot_robustness}()}
}
\concept{plotting functions}
