% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_robustness.R
\name{plot_robustness}
\alias{plot_robustness}
\title{Plotting the results of the robustness analysis}
\usage{
plot_robustness(
  posterior_density,
  hdi_mode,
  hdi_l,
  hdi_u,
  mixed_ci_width,
  mixed_ci_l,
  mixed_ci_u
)
}
\arguments{
\item{robustness_results}{tibble, output of the \code{\link{analysis_robustness}} function}
}
\value{
The function returns a ggplot object.
}
\description{
The function visualize the results of the robustness
analysis \code{\link{analysis_robustness}}.
}
\seealso{
Other plotting functions: 
\code{\link{plot_confirmatory}()},
\code{\link{plot_exploratory}()}
}
\concept{plotting functions}
