% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Function to read and merge data}
\usage{
read_data(type)
}
\arguments{
\item{type}{character, either "test", "pilot" or "live"}
}
\value{
The function returns a raw dataframe that contains all the trials from every
participant.
}
\description{
The test, pilot and live datafiles are stored in the projects Github repository.
This function reads, orders by date and merge the individual datafiles of the
chosen repository.
}
\examples{
\donttest{
tpp_raw_data <- read_data(type = "test")
}
}
