% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_robustness.R
\name{robustness_bf_analysis}
\alias{robustness_bf_analysis}
\title{Robustness test of Bf results with Bayesian parameter estimation}
\usage{
robustness_bf_analysis(success, total_n)
}
\arguments{
\item{success}{numeric, total number of successful guesses on erotic trials}

\item{total_n}{numeric, total number of all erotic trials}
}
\value{
The function returns a list of five, the posterior density, the
mode of the HDI, the upper and lower bound of the HDI and the inference
that was based on these values (see \code{\link{inference_robustness_bf}}).
}
\description{
Robustness of BF results is tested by calculating HDI of the posterior
distribution and checking its relation to the region of practical equivalence (ROPE),
promoted in Kruschke, J. K., & Liddell, T. M. (2017). The Bayesian New Statistics:
Hypothesis testing, estimation, meta-analysis, and power analysis from a Bayesian perspective.
Psychonomic Bulletin & Review, 1-29.
}
