% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_robustness.R
\name{robustness_nhst_analysis}
\alias{robustness_nhst_analysis}
\title{Robustness test of Bf results with NHST}
\usage{
robustness_nhst_analysis(success, total_n)
}
\arguments{
\item{success}{numeric, total number of successful guesses on erotic trials}

\item{total_n}{numeric, total number of all erotic trials}
}
\value{
The function returns a list of three, the p-value of the
equivalence test, the p-value of the equality test, and the inference 
\code{\link{inference_robustness_nhst}} made based on the values
and alpha `r analysis_param$inference_threshold_nhst`.
}
\description{
Robustness of Bf results is tested with NHST proportion tests.
In this function we perform both an equivalence test and
an equality test to draw statistical inference.
}
