% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_descriptives.R
\name{sample_descriptives_current}
\alias{sample_descriptives_current}
\title{Current descriptive}
\usage{
sample_descriptives_current(processed_data)
}
\arguments{
\item{processed_data}{dataframe, containing only the erotic trials}
}
\description{
description
}
