% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculation_helpers.R
\name{smoothie}
\alias{smoothie}
\title{Function to fit smoothing spline}
\usage{
smoothie(x, y)
}
\arguments{
\item{x}{numeric, the x param}

\item{y}{numeric, the y param}
}
\description{
For the confirmatory analysis plot the cumulative
Bayes factors calculated with three priors needs
to be smoothed to ease understanding.
}
