% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{split_data}
\alias{split_data}
\title{Function to split data according to stopping points}
\usage{
split_data(df)
}
\arguments{
\item{df}{dataframe, the input dataframe}
}
\value{
The function returns a dataframe with the following structure:
  \itemize{
  \item checkpoint, integer,the number of investigated erotic trials at a given checkpoint
  \item split_data, df, the input dataframe until with trials from 1 to the given checkpoint
  \item success, integer, the number of successful guesses
  \item total_n, integer, the number of erotic trials
  \item n_iteration, integer, the number of iteration (see \code{\link{confirmatory_mixed_effect}})
  }
}
\description{
This function cuts the processed data at each checkpoint
that is passed and saves the result to a list of dataframes.
The function also calculates the total number of erotic trials,
and the total number of successful guesses for the each checkpoint.
Finally, it saves the number of iterations that is needed for the
correction of alpha for NHST tests. The function can be used
to prepare a dataframe to run the main confirmatory analysis
on all the passed checkpoints iteratively.
}
\examples{
\donttest{
# Including only erotic trials
tpp_processed_data <- clean_data(raw_data = example_m0)
# Splitting data by checkpoints
tpp_split_data <- split_data(df = tpp_processed_data)
}
}
