% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{tell_checkpoint}
\alias{tell_checkpoint}
\title{Function to return checkpoint information}
\usage{
tell_checkpoint(df)
}
\arguments{
\item{df}{dataframe, the input dataframe}
}
\value{
The function returns a list of three. The number of trials,
the index of the currently passed checkpoint, the index of the
next checkpoint.
}
\description{
This function checks the number of trials in the input
dataframe and determines the index of the currently passed
checkpoint and the next checkpoint based on the stopping points
determined in the sequential analysis plan. If raw data are provided
the function drops all the non-erotic trials (see \code{\link{clean_data}}).
The function uses the following checking points: (`r analysis_params$when_to_check`).
}
\examples{
\donttest{
# Including only erotic trials
tpp_processed_data <- clean_data(raw_data = example_m0)
# Get checkpoint
tell_checkpoint(df = tpp_processed_data)
}
}
