% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_text_output_helpers.R
\name{text_helper_current_confirmatory}
\alias{text_helper_current_confirmatory}
\title{Main confirmatory current results descriptive text}
\usage{
text_helper_current_confirmatory(bf_replication, bf_uniform, bf_buj)
}
\description{
Main confirmatory current results descriptive text
}
