% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_text_output_helpers.R
\name{text_helper_current_robustness}
\alias{text_helper_current_robustness}
\title{Robustness current results descriptive text}
\usage{
text_helper_current_robustness(
  inference_confrimatory_bf,
  inference_robustness_bf
)
}
\description{
Robustness current results descriptive text
}
