% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_confirmatory.R
\name{analysis_confirmatory}
\alias{analysis_confirmatory}
\title{Primary confirmatory analysis}
\usage{
analysis_confirmatory(raw_data)
}
\arguments{
\item{raw_data}{dataframe, df containing all trials}
}
\value{
The output is a dataframe containing the inputs, the results
  of the four analysis, and the inferences at each checking point until
  an inference supporting the M0 or M1 model is reached or until the
  last checking point.
  The list of variables in the output dataframe:
  \itemize{
  \item checkpoint, integer,the number of investigated erotic trials at a given checkpoint
  \item split_data, df, the input dataframe until with trials from 1 to the given checkpoint
  \item success, integer, the number of successful guesses
  \item total_n, integer, the number of erotic trials
  \item n_iteration, integer, the number of iteration (see \code{\link{confirmatory_mixed_effect}})
  \item mixed_effect_res, df, the results of the mixed-effect logistic regression
    \item mixed_ci_width, numeric, width of the confidence interval
    \item mixed_ci_lb, numeric, lower limit of the confidence interval 5%
    \item mixed_ci_ub, numeric, upper limit of the confidence interval 95%
  \item bf_res, df, the results of the three Bayes factor analysis
    \item bf_replication, numeric, rounded Bf of analysis with prior based on Bem 2011 experiment 1 results
    \item bf_uniform, numeric, rounded Bf of analysis with uniform prior
    \item bf_buj, numeric, rounded Bf of analysis with BUJ prior
  \item inference, character, the summarized inference
  }
}
\description{
This function conducts the primary confirmatory analysis
at each passed stopping point. In the analysis the mixed-effect logistic
regression (see \code{\link{confirmatory_mixed_effect}}) and three Bayes factor
analysis (see \code{\link{confirmatory_bayes_factor}}) are conducted.
Inferences are made for the four tests separately and then summarized
into the final inference of the primary analsyis.
}
\examples{
\donttest{
# Running the confirmatory analysis
confirmatory_result <- analysis_confirmatory(raw_data = tpp_raw_data)
# Checking the inferences at each checking point
confirmatory_result$inference
}
}
\concept{analysis functions, confirmatory functions}
