% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_exploratory.R
\name{analysis_exploratory}
\alias{analysis_exploratory}
\title{Exploratory analysis}
\usage{
analysis_exploratory(confirmatory_results)
}
\arguments{
\item{confirmatory_results}{tibble, output of the \code{\link{analysis_confirmatory}} function}
}
\value{
The function returns a list of three, the possible success
rates, the the proportion of observed successful guess rate, and the
proportion of proportion of simulated successful guess rate.
}
\description{
This function runs the exploratory analysis that
aims to explore the empirical distribution of successful
guess rate in the population, by contrasting it to a theoretical
distribution expected if the true successful guess chance is homogeneously
50% in the population. This may provide information for future research on
potential irregularities in the distribution of guess rates.
}
\section{Note}{
 Exploratory analysis results will not affect
  the conclusions of our study.
}

\examples{
\donttest{
# Running the primary confirmatory analysis
confirmatory_results <- analysis_confirmatory(raw_data = example_m0)
# Running the exploratory analysis
analysis_exploratory(confirmatory_results = confirmatory_results)
}
}
\concept{analysis functions}
