% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_robustness.R
\name{analysis_robustness}
\alias{analysis_robustness}
\title{Robustness analysis}
\usage{
analysis_robustness(confirmatory_results)
}
\arguments{
\item{confirmatory_results}{tibble, output of the \code{\link{analyis_confirmatory}} function}
}
\value{
The function returns a list containing the results
of the frequentist proportion test and the Bayesian parameter estimation
as nested lists, the final inference of the robustness test as a character
vector, and the results of the ,
}
\description{
Two alternative variants of the Bayesian proportion test in the primary
confirmatory analysis: frequentist proportion test (see \code{\link{robustness_nhst_analysis}}),
and full-Bayesian hypothesis testing using parameter estimation (see \code{\link{robustness_bf_analysis}}).
}
\section{Note}{
 Robustness analysis results will not affect
  the conclusions of our study.
}

\examples{
\donttest{
# Running the primary confirmatory analysis
confirmatory_results <- analysis_confirmatory(raw_data = example_m0)
# Running the robustness analysis
analysis_robustness(confirmatory_result = confirmatory_results)
}
}
