% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_confirmatory.R
\name{cumulative_bayes_factor}
\alias{cumulative_bayes_factor}
\title{Calculating cumulative Bayes factors}
\usage{
cumulative_bayes_factor(df)
}
\arguments{
\item{df}{dataframe, the input dataframe}
}
\value{
The function returns a wide formatted dataframe with
the number of trials, and the Bayes factors with the three different
trials.
}
\description{
This function calculates the cumulative Bayes factors
with three priors. The output of this function is used for
\code{\link{plot_confirmatory}} as an input. The function
can accept a dataframe that is already only containing the
cumulative success variable \code{\link{cumulative_success}},
or a raw dataframe.
}
\examples{
\donttest{
# Calculate the cumulative Bayes factors
cumulative_results <- cumulative_bayes_factor(df = example_m0)
}
}
