% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_confirmatory.R
\name{plot_confirmatory}
\alias{plot_confirmatory}
\title{Plotting the results of the primary confirmatory analysis}
\usage{
plot_confirmatory(cumulative_results, animated = FALSE)
}
\arguments{
\item{cumulative_results}{dataframe, output of the \code{\link{cumulative_bayes_factor}} function}

\item{animated}{logical, whether the plot to be animated with plotly or not}
}
\value{
The function returns a ggplot object.
}
\description{
The function visualize the results of the confirmatory
Bayes factor analysis \code{\link{cumulative_bayes_factor}}.
The Bayes factors calculated at each new experimental trial
are displayed on this figure.
}
\seealso{
Other plotting functions: 
\code{\link{plot_exploratory}()},
\code{\link{plot_robustness}()}
}
\concept{plotting functions}
