% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exploratory.R
\name{plot_exploratory}
\alias{plot_exploratory}
\title{Plotting the results of the exploratory analysis}
\usage{
plot_exploratory(exploratory_results)
}
\arguments{
\item{exploratory_results}{tibble, output of the \code{\link{analyis_exploratory}} function}
}
\value{
The function returns a ggplot object.
}
\description{
The function visualize the results of the exploratory
analysis \code{\link{analyis_exploratory}}.
}
\seealso{
Other plotting functions: 
\code{\link{plot_confirmatory}()},
\code{\link{plot_robustness}()}
}
\concept{plotting functions}
