% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_text_output_helpers.R
\name{text_helper_current_general}
\alias{text_helper_current_general}
\title{General current descriptive text}
\usage{
text_helper_current_general(sample_descriptives_current_res, robustness_bf_res)
}
\description{
This function creates a character string for
the Shiny app. The text describes the general
descriptive information about the study.
}
